ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pydantic==2.6.0
python-multipart==0.0.9
qdrant-client==1.7.3
google-generativeai==0.3.2
pdfplumber==0.10.3
python-docx==1.1.0
requests==2.31.0
google-genai
fastembed